don't go into infinite loop when prepping items
This code was meant to randomly choose items from
each section, which is now done server-side.
It could still be improved to do that again for
xml fetched from a different server, but this implementation
would go into an infinite loop if the qti had bad ids
for some items.