Make student selections show in quiz-attempt summary
It can be helpful for a teacher to be reminded of the
correct answers and to see which incorrect answer the
student chose.
The incorrect answer may be an indicator of a concept
that the student needs additional help with.

Test Plan:

-Go to course with teacher role and click "Quiz 1"
-Click "Manage Quiz Attempts"
-Click on a quiz-attempt
-Notice the green checks showing the correct answers and
the red x's marking incorrect answers the student chose.