Return new xml of questions answered
If there is access to an xml document that
contains only the questions that were randomly
chosen for a student quiz, that xml can
be used to reference that specific quiz instance
later on. A good use of this is for students to
view their past quiz attempts.

Test Plan:

run these commands in terminal:

-bundle exec rspec /Users/juliannagreen/Desktop/OpenAssessments/spec/models/assessment_xml_spec.rb:20

-bundle exec rspec /Users/juliannagreen/Desktop/OpenAssessments/spec/models/assessment_result_spec.rb:105