refactor LTI outcome code to Lti::AssessmentResultReporter
This was all inside the AssessmentResult itself which was
a bit messy. This allows the AR to not know how to report
LTI outcome scores.

This is also the first step in updating the outcome behaviour
to use the info stored on an LtiLaunch instead of making the
AssessmentResult store that info itself, and using the
LtiCredentials used at launch time, not just the account
default.