BUG: dropdown questions showing correct answer
When a student takes a formative/practice assessment, and they
select the incorrect answer in a multi-dropdown question, and they
select/click a confidence button, the dropdown changes to be the
correct answer while the feedback message shows they got the
question incorrect.

This is not desired behavior and is confusing to the student.