Iron out base webpack/babel config
Starting from scratch with the most recent versions of webpack and
babel was the easiest way to get the base app to compile.

This commit also reintroduces the missing actions, stores, models,
themes, and utils.

So at this point the base app is compiling with webpack and babel, but
next we need to npm install the newest versions of the component
dependencies and ensure their appropriate functionality.